+++ src/ops/base/destroy/scripts/remove_mkcert.sh
@@ -1,17 +1,15 @@
 #!/bin/bash
 set -euo pipefail
 
-DOMAIN=$1
-
 # Remove certs
-sudo rm -f /etc/nya/certs/${DOMAIN}.pem
-sudo rm -f /etc/nya/certs/${DOMAIN}-key.pem
+sudo rm -f /etc/nya/certs/{{ domain }}.pem
+sudo rm -f /etc/nya/certs/{{ domain }}-key.pem
 
 # Remove mkcert CA
-mkcert -uninstall
+mkcert -uninstall || true
 
 # Remove mkcert binary
 sudo rm -f /usr/local/bin/mkcert
 
 # Remove certs directory if empty
-sudo rmdir --ignore-fail-on-non-empty /etc/nya/certs
+[ -d /etc/nya/certs ] && sudo rmdir --ignore-fail-on-non-empty /etc/nya/certs || true
